Keep shortcut regression coverage below file budgets
The Ctrl+P remap/unbind fix added runtime coverage and helper logic to files that are already tracked by the Swift file-length guard. Move the new coverage and lookup/routing helpers into small focused files so CI can enforce the existing budget without accepting new long-file debt.

Constraint: workflow-guard-tests rejects growth in tracked Swift files over the budget

Rejected: Refresh the Swift file-length budget | would accept new long-file debt for a localized shortcut fix

Confidence: high

Scope-risk: narrow

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv

Tested: jq empty Resources/Localizable.xcstrings web/messages/en.json web/messages/ja.json web/data/cmux.schema.json web/data/cmux-settings.schema.json

Tested: bash -n scripts/build-ghostty-cli-helper.sh scripts/reload.sh

Tested: plutil -lint GhosttyTabs.xcodeproj/project.pbxproj

Not-tested: XCTest execution; repository policy keeps tests in CI
