+++ Sources/KeyboardShortcutSettingsLookup.swift
@@ -1,0 +1,45 @@
+import Foundation
+
+extension KeyboardShortcutSettings {
+    static func shortcutIfBound(for action: Action) -> StoredShortcut? {
+        #if DEBUG
+        shortcutLookupObserver?(action)
+        #endif
+
+        if let managedShortcut = settingsFileStore.override(for: action) {
+            return managedShortcut.isUnbound ? nil : managedShortcut
+        }
+
+        guard let data = UserDefaults.standard.data(forKey: action.defaultsKey),
+              let shortcut = try? JSONDecoder().decode(StoredShortcut.self, from: data) else {
+            let defaultShortcut = action.defaultShortcut
+            return defaultShortcut.isUnbound ? nil : defaultShortcut
+        }
+        return shortcut.isUnbound ? nil : shortcut
+    }
+
+    static func shortcut(for action: Action) -> StoredShortcut {
+        shortcutIfBound(for: action) ?? .unbound
+    }
+
+    static func menuShortcut(for action: Action) -> StoredShortcut {
+        guard !KeyboardShortcutRecorderActivity.isAnyRecorderActive else {
+            return .unbound
+        }
+        return shortcut(for: action)
+    }
+
+    static func isManagedBySettingsFile(_ action: Action) -> Bool {
+        settingsFileStore.isManagedByFile(action)
+    }
+
+    static func unbindShortcut(for action: Action) {
+        setShortcut(.unbound, for: action)
+    }
+
+    static func settingsFileManagedSubtitle(for action: Action) -> String? {
+        guard isManagedBySettingsFile(action) else { return nil }
+        return String(localized: "settings.shortcuts.managedByFile", defaultValue: "Managed in cmux.json")
+    }
+
+}
